+++ lib/hcloud/entities/server_type_location.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module HCloud
+  class ServerTypeLocation < Entity
+    attribute :id, :integer
+    attribute :name
+    attribute :deprecation, :deprecation
+  end
+end